+++ packages/spacecat-shared-utils/src/url-helpers.js
@@ -10,6 +10,12 @@
  * governing permissions and limitations under the License.
  */
 
+import { context as h2, h1 } from '@adobe/fetch';
+
+/* c8 ignore next 3 */
+export const { fetch } = process.env.HELIX_FETCH_FORCE_HTTP1
+  ? h1()
+  : h2();
 /**
  * Prepends 'https://' schema to the URL if it's not already present.
  * @param {string} url - The URL to modify.
@@ -72,8 +78,16 @@ function composeBaseURL(domain) {
   return baseURL;
 }
 
+async function composeAuditURL(url) {
+  const urlWithScheme = prependSchema(url);
+  const resp = await fetch(urlWithScheme);
+  const finalUrl = resp.url.split('://')[1];
+  return stripTrailingSlash(finalUrl);
+}
+
 export {
   composeBaseURL,
+  composeAuditURL,
   prependSchema,
   stripPort,
   stripTrailingDot,